+++ dectest/src/backend/decimal32.rs
@@ -14,6 +14,7 @@
 // limitations under the License
 
 use std::cmp::Ordering;
+use std::hash::Hasher;
 use std::str::FromStr;
 
 use dec::{Class, Context, Decimal128, Decimal32, Decimal64, Rounding, Status};
@@ -27,6 +28,7 @@ pub struct Decimal32Backend {
 impl Backend for Decimal32Backend {
     type D = Decimal32;
 
+    const HASHABLE: bool = false;
     const REPORTS_STATUS_CLAMPED: bool = false;
     const REPORTS_STATUS_ROUNDED: bool = false;
     const REPORTS_STATUS_SUBNORMAL: bool = false;
@@ -71,6 +73,13 @@ impl Backend for Decimal32Backend {
 
     fn to_decimal128(&mut self, n: &Self::D) -> Decimal128 {
         Decimal128::from(*n)
+    }
+
+    fn hash<H>(_: &Self::D, _: &mut H)
+    where
+        H: Hasher,
+    {
+        unimplemented!("Decimal32 does not implement Hash")
     }
 
     fn status(&self) -> Status {